HPCC-33113 Ensure empty compressed part size is correct
When writing a logical file that is wider than Thor with,
Thor pads the file at publish time with empty parts.
If compressed it creates an empty compressed file, which
will typically be non-zero in size (e.g. contains compressed
header). The logical file part meta data did not relect this
correct physical size.
Consequently, as a result of the check added in HPCC-33064,
reading these files was provoking this error check.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>