HPCC-33150 Avoid padding logical files with empty parts
When targeting a cluster/plane that was wider than the Thor
instance, Thor used to ensure the target file width matched the
width of the target cluster (or plane), by in part, creating
empty physical files.
This behaviour becomes more common in a containerized world,
where the target plane is normally the default data plane with
a width that doesn't relate to a Thor instance.

Turn off this behaviour by default, allow it to be re-enabled
in configuration or on a per output basis as a fail-safe.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>