HPCC-33137 Add supports for executing the cost optimizers in Thor
Cost optimizer will execute in Thor for containerized deployments.  It will continue to execute
in EclAgent for non-containerized deployment.

The reason for this change is that in containerized deployments EclAgent cannot calculate the
cost of the Thor cluster and so cannot not assigned the cost of issues reported by the cost
optimizer.  It cannot calculate the cost of the Thor cluster because Thor's cost parameters
and resource configuration is not available to EclAgent - this information is only available
in the Thor's configuration.

Note, that as the Thor manager executes on a per job basis, it executes the analyzer after
every graph executes.  The concept of "end of workunit" does not exist at present in Thor so
the analyzeWhenComplete==true has not been implemented for Thor.

The changes:

* Support for executing cost optimizer in Thor.  This the default in containerized.
* New 'analyzeInEclAgent' option - if true, the analyzer executes in EclAgent, otherwise
it executes in Thor Manager.
* By default 'analyzeInEclAgent' is true in bare-metal and false in containerized
* The defaults for 'analyzeWhenComplete' has changed for containerized.  It is now false by
default in containerized.  It remains true in bare-metal.
* New 'disabled' option is available within analyzerOptions to disable cost optimizer.  'disabled'
is false by default.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>