HPCC-33356 Capture post-mortem logging on job failure
This extends capture of the verbose post-mortem logging to an
[compressed] archive to occur (optionally) whenever a workunit
fails.
NB: if there is a subsequent follow on error and the process or
container exits badly, the managing scripts will also capture
and associate post mortem logging with the workunit.

By default, all workunits that fail will capture post mortem
logging and associate it with the workunit.
It can be configured at values yaml level, or at the workunit
level, e.g. with:

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>