HPCC-32989 Fix potential stall in spilling distributors
Distributors spill if their receiving thread starts to
receive too much (e.g. if unbalanced).
The new lookahead implementation (CCompressedSpillingRowStream)
is conditionally used by distributors in this situation
[dependent on #option('newlookahead') ].
If a spill occurs, and the consumers last content is from
disk (as opposed to from the queue), then it fails to signal
the completion correctly.
If the provider/writer to CCompressedSpillingRowStream calls
flush after this event, it will block waiting for the consumer
to signal that is complete, and stall indefinitely.

NB: this could happen in any context a distributor is used,
e.g. regular hash distribute, but only if there is
memory pressure whilst it's actively reading.
A smart join that has failed over to a local (hash distributed)
smart join is a common case where that can happen.

NB2: standard lookahead that also uses CCompressedSpillingRowStream
inside CRowStreamLookAhead, is unaffected, due to the way it
asynchronously ensures stop() is explicitly called (which signals
a blocked flush to release).

NB3: although the bug is in CCompressedSpillingRowStream introduced
in 9.6.26 (HPCC-32132), in practice this should only affect
distributors, and distributors only started to use this new
spilling container (via option, and not in by default in BM) since
9.8.16 (HPCC-32486).

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>