HPCC-33381 Fix dfuaccess crash regression
HPCC-33341 upped the DAFILESRV_METAINFOVERSION, but not the
rowservice server side handling of the new version number, causing
it to crash.
For all intents and purposes the deserialization format is the same
except for the addition of extra meta info ('group') within the
metaInfo property tree.
Detect and handle the new version, and improve the error if an
invalid version is detected.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>