HPCC-33635 Optimize updates to file properties following disk reads
Updates to numDiskReads and readCost by thor and hthor have been modified
to improve performance:
* The numDiskReads and readCosts updates to a file's properties are made
once per activity.
* When a subfile is read, both the superfile's properties and the subfile's
properties are updated.  However, where a subfile being updated belong to
other superfiles (in addition to the superfile being read), then other
owning superfile's properties will not be updated.  The reason for not
updating all owning superfiles is to avoid negatively impacting performance.
That does mean that where a subfile belongs to more than one superfile, the
superfile properties stats will not always be an aggregate of the stats in
its subfile.
* Updates to file properties are bypassed, if numDiskReads and readCost are 0.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>