HPCC-33811 Maintain compatibility with version 1.01
Adjust the jsonpath result from the kubectl call to have the same format as the
raw result. This allows the current client code to continue to run on 1.01, and
still allow an update to use the structured response starting on v1.02.

This change prompted by the failure of the hpcc4j regression test.

Signed-off-by: Terrence Asselin <terrence.asselin+copilot@lexisnexisrisk.com>