HPCC-33928 Fix various striping/dirPerPart issues
1) Never stripe or add dir-per-part dirs, if parts have override
2) Do not remove dir-per-part meta flags when creating super file
descriptor (was done to avoid workers re-adding, but now done via [1]).
3) Ensure CDistributedFilePart::getPartDirectory stripes consistently.
It was not doing so in general, and relied on the workers doing so
(now not done due to [1]). This had the effect of when used of
the directory being wrong, and when used to build a super
IFileDescriptor the overridenames were wrong, but were patched by
the worker code.
Now with the fixes, CDistributedFilePart::getPartDirectory will return
a fully striped/dir-per-part dir, and the workers will not manipulate
further if used as part of an overridename.
4) Tighen up the cluster/group lookups when a foreign file is involved.
Avoid situations where the client environment lookup a group
name from a foreign environment and spuriously matched a local group.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>