HPCC-34181 Instructions and configs for LDAP in development
Add developer documentation and supporting configuration files to run a basic
platform deployment (bare-metal or k8s) with LDAP auth on ECL Watch.

Instructions include running a 389ds LDAP Directory Server in a docker container
with a persistent volume for the server state. A web-based LDAP admin is
included in the container to troubleshoot and make any advanced changes needed.

The platform config has minimal changes from the basic configuration supplied at
install.

The scope of this work is to make it easier for developers, QA and other technical
platform users to test, create and troubleshoot features that require LDAP
authentication. It is in no way designed to be used in any production environment
or in any application that contains sensitive data.

Signed-off-by: Terrence Asselin <terrence.asselin+copilot@lexisnexisrisk.com>