HPCC-34200 ECL Watch report failure of fetch Results
When attmpting to fetch results for a given workunit, it is possible
that ESP will throw and catch an exception, and put the message from
that exception into a "ResultsDesc" property of the WUInfoResponse.
This changes both ECL Watch v5 & v9 to look for this property when
fetching results and to display an error message popup, as appropriate.

Signed-off-by: Jeremy Clements <79224539+jeclrsg@users.noreply.github.com>