HPCC-34432 Fix dafilesrv file secret mapping bug
When a secure remote (~remote::) logical file is accessed, the
endpoints of the parts (directio service) are mapped to the
secret associated with the remote storage definition,
and unmapped when the logical file is no longer used.

Multiple logical remote logical files may map in this way,
and therefore the mapping is link counted.
However, there was a bug when adding, that it did not increment
the link count, consequently, when the 1st logical file unwound,
it deleted the secret.

This could cause current active files, that were yet to connect,
to fail with "Failure to establish secure connection", which may
be caught and ultimately seen as:
"Failed to connect (setting host down) to dafilesrv/daliservix"

And, a secondary consequence was that the subsequent logical files
that unwound and tried to delete their mapping, hit an assert in
a destructor, causing the process to abort.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>