HPCC-34429 Calc wf and global cost from children and remove CostExecute from StatisticsAggregator list
Changes:
- aggregateCost function modified to always calculate the cost from the child scopes, rather than
was using any values already in the global stats.
- Remove the StCostExecute from StatisticsAggregator list of stats to aggregate. (The reason for
doing this was that the StCostExecute was being written after the StatisticsAggregator was called,
so the StatisticsAggregator didn't USUALLY have subgraph StCostExecute to aggregate.  So,
StCostExecute was only included in the aggregate stats when the StatisticsAggregator executed when
the next subgraph executed. Of course, that meant that the last subgraph stats were never included
in the aggregated cost values.  Also, the graph, workflow and global costs cannot be accurately
calculated by summing subgraph costs, as there are gaps between the execution of subgraphs)

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexisrisk.com>