HPCC-34418 ecl unused-files --check-packagemaps should match documentation
The ecl roxie unused-files --check-packagemaps command was excluding (counting
as 'used') any file referenced in _any_ packagemap even if the packagemap
was disabled. This is contrary to what the usage and documentation states.

With this change, a file is counted as 'used' for the --check-packagemaps
option when it wouldn't have been otherwise only if it is in an active
packagemap.

Put another way, files referenced only in packagemaps are unused only if
the packagemap is disabled.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>