HPCC-34486 Refactor DeMonServer to make it easier to understand and more logical
- The statistic gathering functionality has been moved to a new class to separate the
gathering functionality from the DeMonServer processes.
- A new startGraph function has been created to replace the use of the reportGraph.  reportGraph
was always gathering stats from the graph but it was not necessary to do this when the
graph has just started.
- The reportGraph function was always writing the WhenStarted time stamp.  This is only
necessary when the graph has just started so this has been moved to the startGraph.
- wfid and graphName are now passed to reportStatus as it is always ready and available
when the function is called. The code in reportStatus to query wfid and graphName has been
removed.
- Some variables (such as graphName) have been camel cased.
- Some variables have been renamed to make them more descriptive.
- graphStartMsTick (previously graphStart) is killed when the endGraph is called.  This is
necessary as errors or aborts were leaving the graphStartMsTick with old values.
- There is one small functional change: if the graph has completed, then the guillotine limit
is not checked.  (It is not necessary to kill a job if it has already completed.)
- There is no other functional change.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexisrisk.com>