HPCC-34528 Avoid eclcc k8sjob trying to launch another k8s job
If eclcc fails within a eclccserver k8s job, the logic was
treating it as if it had timedout, and caused it to try to run
another k8s job inside the 1st. Since the eclccserver k8s job
has a service account with no rights to access the kube API,
it would fail with a kube API access error.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>