HPCC-34545 Fix BM writing to PLANE
Writing to a PLANE, did not pickup the correct prefix.
One of the consequences of that was that when the plane was
configured with Azure Blob API details, the correct prefix was
not injected into the file descriptor (as the directory prefix),
and consequently cause the file to be written into the wrong
place (was written locally).

NB: containerized did work correctly. This PR now uses the same
code path as containerized (via getDataStoragePlane()).

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>