HPCC-34558 Remove linking of userDesc in CDFSFile::queryUserDescriptor
The only place that CDFSFile::queryUserDescriptor() is used is in CServiceDistributedFileBase::CDistributedSuperFileIterator::setCurrent
member function. This function uses source->queryUserDescriptor() as an argument in the call to lookupDFSFile:

    void setCurrent(unsigned w)
    {
        VStringBuffer lfn("~remote::%s::%s", source->queryRemoteName(), owners[w].c_str());
        Owned<IDFSFile> dfsFile = lookupDFSFile(lfn, accessMode, source->queryTimeoutSecs(), keepAliveExpiryFrequency, source->queryUserDescriptor());
	...

In lookupDFSFile, the userDesc is used as follows,
- calls getDfsClient(.., userDesc):  getDfsClient queries username and password from userDesc.  It doesn't doesn't own or link userDesc
or pass it to any function or object.
- calls ensureClientLease(.., .., .., userDesc): ensureClient doesn't use userDesc at all.
- calls createDFSFile(.., .., .., .., userDesc): createDFSFile creates a CDFSFile object using userDesc as an argument in the constructor.
The constructor initializes the member variable userDesc.  The member variable userDesc is Linked<IUserDescriptor>.  Linked will mean that
userDesc  link counted.

CDFSFile::queryUserDescriptor() is not used in any other place.  CDFSFile::queryUserDescriptor() shouldn't return a linked pointer.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexisrisk.com>