HPCC-24831 ECL Watch improve v9 File protected by state
Changes the ECL Watch v9 File Details page to use a button to set and
unset the file's protected state for the current user. Also fixes an
issue where the "Protected By" tab's count was not updating properly.

This PR also includes the change to LogicalFileSummary.tsx from
HPCC-34159, made to IndexFileSummary.tsx. It should've been included
then, but was missed.

Signed-off-by: Jeremy Clements <79224539+jeclrsg@users.noreply.github.com>