HPCC-34267 Support to view containerized eclccserver config in ECL Watch
- Update dalikvstore with a class to easily set/get component configs in store.
- Update eclccserver and esp to store configs on startup.
- Create WsTopology method to get list of available component config names.
- Create WsTopology method to retrieve component config by name.

Configs are stored/retrieved in xml/yaml depending on deployment type. XML for
bare metal and YAML for containerized. A synthetic "global" component is recognized
for getting the global config.

Some changes made to the dalikvstore files were to workaround compiler errors, such
as commenting a pragma, changing a size type and moving some const defns.

This is a WIP submitted for feedback.

Signed-off-by: Terrence Asselin <terrence.asselin+copilot@lexisnexisrisk.com>