HPCC-34172 owner costs, avoid super lookup for every sub
Use a IDistributedFileTransaction in
FileReadPropertiesUpdater::publish(), to avoid pathological cost
of re-looking up the owning super file for every subfile read.

In practice this meant that when iterating through owning supers,
it looked the super up, and all the subfiles (include all locking).

i.e. if a super file with 10 subfiles was read, publish() would:
1) find the super (getOwningSuperFiles)
2) look them up each (just 1 in this example) super via DFS.
This would involve lookup all (10) subfiles, including locking.
3) update cost in super
* Repeat for all other subfiles.

Employing a IDistributedFileTransaction, will cause the 2nd
iteration to find any super that has previously been found as
an owner, to be found within the transaction rather than
re-looking up.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>