HPCC-34575 Patch XSS injection of WsSMC/Activity ChatURL
Encode ChatURL value as per OWASP recommendation before returning. This
prevents it from being executed as code when assigned to a JS variable, but
does not prevent its use as a link elsewhere on the page.

Recommended output encoding for JS contexts is to use the unicode `\uXXXX`
hex code point format for all non-alphanumeric characters.

Continue to strip out all nonprinting characaters as before.

Signed-off-by: Terrence Asselin <terrence.asselin+copilot@lexisnexisrisk.com>