HPCC-34267 Static config API
Adds an API to WsTopology to retrive static configs. Currently supports
containerized deployments. The ECL Watch pod mounts the ConfigMap volumes
for the other configured components.

TpConfiguredComponents returns a list of all configured and available
components whose configs can be returned. It also includes the "global"
component for the global section of the config.

TpComponentConfiguration takes a list of components, including the special
names "global" and "all". It then returns the serialized configuration for
each component requested.

Signed-off-by: Terrence Asselin <terrence.asselin+copilot@lexisnexisrisk.com>