HPCC-34268 Mask sensitive values in configurations
ESP provides three methods of retrieving configuration files, two of which are
available to non-admin users. This change fully masks the sensitive values
included in those two methods:

 - `EspHttpBinding::onGetConfig` handler, invoked using the `config_` URL
   parameter. This is used in the ECL Watch "Configuration" menu, and on the
   "Configuration" tab of the "Topology" page of containerized deployments.
 - `CWsTopologyEx::onTpGetComponentFile` function, used on the "Configuration"
    tabs of individually selected components on the "Operations" page.

The masking engine is owned and initialized by the ESPContainer to be available
to both the http binding and service classes. We're initially using a static
config for the masking profile for simplicity, to avoid missing configuration
or mis-configuration, and because these components have a constrained set of
sensitive values.

Future extended use of masking could bring us to consider making changes
including:

 - Swapping out the current masking plugin for an efficient implementation
   that evaluates rules in parallel. Testing shows about 70% increase in
   latency when masking, but it still has sub-millisecond responses.
 - Additional profiles for masking data of different types and formats.
 - Moving the configuration out of code for flexibility.

Signed-off-by: Terrence Asselin <terrence.asselin+copilot@lexisnexisrisk.com>