HPCC-35372 Fix transaction slowdown regression
The fix for HPCC-34374 inadvertently introduce a significant
slowdown. Changes in the CDeltaWriter thread should have released
the pendingCrit CS before saving, but held it, that meant all
other transactons queuing transactions were held up whenever the
CDeltaWriter was writing.

This was not that noticable in most environments (BM), but in
Azure where the Dali storage is backed by remote storage, it caused
a significant slowdown.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>