HPCC-35314 Filtered index read regression (fail on BM)
HPCC-33751 introduced changes which tried to get the block size
for the index on the Thor manager using findPlaneAttrFromPath.
On a cloud system with local storage, that succeeds as all hosts
in the plane are "localhost", however, on BM (or planes backed by
hosts), that is not the case, and this called failed.

The new code should instead get the plane name directly from the
IDistributeFile it's dealing with.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>