HPCC-35410 Move "Authentication failed:" message info to TxSummary
Remove the high frequency messages beginning "Authentication failed:", as this
information is already included in the TxSummary.

Add to the summary an `authSession` flag instead, indicating if the ESP sent a
Basic Auth Challenge (for non-session requests) or redirect to a login page
(for session requests).

This preserves possibly useful information about the kind of request that
failed auth but in a much more terse form.

Signed-off-by: Terrence Asselin <terrence.asselin+copilot@lexisnexisrisk.com>