HPCC-31368 Fix containerized issue with KJ and out-of-cluster parts
(see when using ~foreign).

Containerized versions treated all file parts as being 'local to
the cluster', which meant they were striped across workers, each
dealing with a slice of them. Each worker would send requests to
parts in other slices using remote handlers.
However, the remote handler code assumed the index parts were
local (as most parts are in a containerized setup), this assumption
broke down for out-of-cluster files, e.g. foreign indexes used in
a KJ. The consequence was that KJ would fail with:
"Failed to open index file".

The fix is 3 fold:
1) Add logic to check if parts are on localhost (mounted), and
stripe if they are (this mirrors the previous behaviour where
all were, and mirrors what BM is trying to acheive where 'local' parts
are striped across their corresponding cluster hosts).

2) Add an option 'keyedJoinStripeOutOfClusterLookups' that forces
all parts to be striped (default off for now). i.e. this will cause
out-of-cluster parts to be striped across workers too.
t sm
3) Change the remote handlers to send the full path in the remote
requests (not just local path). Previously remote handlers were
only intended to be used when parts were local to the remote lookup
service. Now with keyedJoinStripeOutOfClusterLookups=true, that might
not be the case.
e.g. if striped and foreign, remote handler requests will be sent
to other workers to lookup parts which are non local. Therefore the
full remote path is needed not just the local path.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>