HPCC-35546 Change storage plane -> dali group update semanics
Ensure Dali groups are kept in sync with their corresponding
storage/planes definitions.
Delete stale old groups that no longer have a corresponding plane

NB: In BM, non-dropzone groups are protected by default, to
maintain legacy behaviour, all non-protected groups are deleted.
However, this would delete all existing groups on the 1st run with
this change. To prevent this, an initial scan of all groups is
performed. If no protected groups are found, it is deemed as the 1st
run with this mechanism enabled, and all existing non-dropzone groups
are set to protected.

Also unify initClusterGroups and initClusterAndStoragePlaneGroups

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>